+++ src/components/BlocklyApp.jsx
@@ -12,21 +12,43 @@ import 'blockly/javascript_compressed';
 import 'blockly/localize/es';
 import 'blockly/components';
 
+import executorGenerator, { type Executor } from 'blockly/executorGenerator';
+import BlocklyToolbox, { type BlocklyToolboxElement } from 'components/BlocklyToolbox';
 import ExecuteButton from 'components/ExecuteButton';
-import BlocklyToolbox from 'components/BlocklyToolbox';
+import { blockNames } from 'blockly/constants';
 
-const id = 'blockly-app';
+const ID = 'blockly-app';
+
+export type GameMetadata = {|
+  blockDefinitions: Array<BlockDefinition>,
+  defaultElements: string,
+  elements: Array<BlocklyToolboxElement>,
+|};
+
+type BlockDefinition = {|
+  name: string,
+  blockExecutor: (...args: Array<*>) => void,
+|};
+type Props = {|
+  gameMetadata: GameMetadata,
+|};
 
 export default class BlocklyApp extends React.Component {
+  props: Props;
+
   handleWorkspaceCreation: (toolbox: HTMLElement) => void;
   handleClick: () => void;
   workspace: Object;
+  executor: Executor;
 
-  constructor() {
-    super();
+  constructor(props: Props) {
+    super(props);
 
-    this.handleWorkspaceCreation = this.handleWorkspaceCreation.bind(this);
-    this.handleClick = this.handleClick.bind(this);
+    this.executor = executorGenerator();
+
+    props.gameMetadata.blockDefinitions.forEach(({ name, blockExecutor }) => {
+      this.executor.addBlockExecutor(name, blockExecutor);
+    });
   }
 
   /**
@@ -36,8 +58,8 @@ export default class BlocklyApp extends React.Component {
    * @param {HTMLElement} toolbox   elements to be used on an exercise
    * @returns {void}
    */
-  handleWorkspaceCreation(toolbox: HTMLElement) {
-    this.workspace = Blockly.inject(id, { toolbox });
+  handleWorkspaceCreation = (toolbox: HTMLElement) => {
+    this.workspace = Blockly.inject(ID, { toolbox });
 
     // 1. Will make orphans a little transparent and they won't be
     //    executed even when Blockly ask to parse workspaceToCode
@@ -45,21 +67,34 @@ export default class BlocklyApp extends React.Component {
 
     // 2. Inject the main block into the workspace
     Blockly.Xml.domToWorkspace(
+      Blockly.Xml.textToDom(`
+        <xml>
+          <block type="${blockNames.ACTION_CONTAINER}">
+            <statement name="program">
+              ${this.props.gameMetadata.defaultElements}
+            </statement>
+          </block>
+        </xml>
+      `),
       this.workspace,
-      Blockly.Xml.textToDom('<xml><block type="ActionContainer" /></xml>'),
     );
   }
 
-  handleClick() {
+  handleClick = () => {
     const code = Blockly.JavaScript.workspaceToCode(this.workspace);
 
-    eval(code);
+    this.executor.run(code);
   }
   render() {
+    const { gameMetadata } = this.props;
+
     return (
       <div>
-        <div id={id} style={{ height: '480px', width: '600px' }}>
-          <BlocklyToolbox handleWorkspaceCreation={this.handleWorkspaceCreation} />
+        <div id={ID} style={{ height: '420px', width: '600px' }}>
+          <BlocklyToolbox
+            elements={gameMetadata.elements}
+            handleWorkspaceCreation={this.handleWorkspaceCreation}
+          />
         </div>
         <ExecuteButton handleClick={this.handleClick} />
       </div>